dockerfile setup for notebook v1
I added a dockerfile for starting up a docker image of a light version of python. it copies the requirements.txt and used pip to install them. It also copies out momentum_notebook into the container. lastly it runs the notebook on a jupyter host and gives out a link through which the user can access the notebook