latex report strawman added & dockerfile update
I added a sample tex file for the latex report together with the bib and logo for title page design.
I adjusted the dockerfile to copy into the container the respective files + a command to run the pdflatex. I also adjusted the readme file with a fifth command in order to mount the resulting pdf to the local host directory